fix: lazy load stdio transport and normalize flags
